fix(#304): Codex round 4 — selector fix (uint16 → uint256) for getEffectiveGraceSeconds
The minimal inline `GRACE_SECONDS_ABI` declared `durationDays`
as `uint16`, but the on-chain `ConfigFacet.getEffectiveGraceSeconds`
takes `uint256` (per `contracts/src/facets/ConfigFacet.sol:1969-1971`).
The 4-byte selector derived from `uint16` would have missed the
on-chain function entirely, falling back to the
`grace_period_end = 0` sentinel and silently negating the
round-3 "live grace schedule" fix.

Selector match: `uint256` parameter type only. Call site updated
to pass `BigInt(durationDays)` matching viem's uint256 expectation.

+++ apps/indexer/src/chainIndexer.ts
@@ -1876,7 +1876,12 @@ const GRACE_SECONDS_ABI = [
     type: 'function',
     name: 'getEffectiveGraceSeconds',
     stateMutability: 'view',
-    inputs: [{ name: 'durationDays', type: 'uint16' }],
+    // On-chain signature is `uint256 durationDays` (per
+    // ConfigFacet.sol:1969). A mismatch here would compile a
+    // different 4-byte selector and the RPC would 404 silently,
+    // falling back to the `graceEnd = 0` sentinel. Codex P2
+    // round-4 on PR #304.
+    inputs: [{ name: 'durationDays', type: 'uint256' }],
     outputs: [{ name: '', type: 'uint256' }],
   },
 ] as const;
@@ -1909,7 +1914,7 @@ async function _resolveGraceEnd(
         address: diamond,
         abi: GRACE_SECONDS_ABI,
         functionName: 'getEffectiveGraceSeconds',
-        args: [durationDays],
+        args: [BigInt(durationDays)],
       })) as bigint,
     );
     return startTime + durationDays * 86_400 + graceSeconds;
